+++ api/libs/helper.py
@@ -113,16 +113,6 @@ def _get_float(value):
     except (TypeError, ValueError):
         raise ValueError('{0} is not a valid float'.format(value))
 
-
-def supported_language(lang):
-    if lang in ['en-US', 'zh-Hans']:
-        return lang
-
-    error = ('{lang} is not a valid language.'
-             .format(lang=lang))
-    raise ValueError(error)
-
-
 def timezone(timezone_string):
     if timezone_string and timezone_string in available_timezones():
         return timezone_string
